Split countdown list view into separate file
