+++ Shared/Styles/OpacityButtonStyle.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct OpacityButtonStyle: ButtonStyle {
+    @Environment(\.isEnabled) private var isEnabled
+
+    func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .opacity(isEnabled ? (configuration.isPressed ? 0.7 : 1) : 0.4)
+    }
+}
+
